PCHR-1550: Fix My Details views to show contact details even if it doesn't have a contract
The contextual filter for the contract period start date was filtering out records without a
start date (that is, contacts without a contract), making the My Details tab empty for such
contracts.

This fixes the issue by allowing an empty start date. In that case, only the information
related to the contact (Name, Work email and Work Phone) will be displayed and the
contract related fields will be empty.