PCHR-1719: Redirect logged-in users from the main path (/) to their accessible dashboard
if the user accessed the site main path e.g www.site-url.com/ or www.site-url/index.php then they should be redirected to
either civicrm T&A dashboard if they had (access CiviCRM) permission or to SSP dashboard in case they don't have this permission.

we know that we are on the main site path by checking if the response of request_path() is empty string .

also since user_acces() function is not accessible inside hook_boot which is needed to check the user permission, I had
to move the call for _user_redirection() function which handles redirection to hook_init.