PCHR-1712: Improve Administration of Content on Welcome Page
Content in welcome page holding logo, site name and slogan was being included
using a custom module with the raw HTML left to be edited to update any
information, which is not ideal if it is supposed to be administered by regular
users.

Removed custom block and added separte blocks with administrable drupal site
logo, site name and slogan to the page.