PCHR-1764: Get Contact Details by Location Name
E-mail and phone data shown on My Details view and staff directory view was
being queried using ID's for location, that were hard-coded on the views. Fixed
by creating CiviHR entity on a databse view that queries work and home e-mail
and phone data using location name instead of ID, and using this entity on
views.