PCHR-2073: Remove Reference to Specific Manager from E-mail
E-mail sent on absence request was being sent to only one manager, even if
contact had several leave approvers configured.  Now, the e-mail sent by rule
on absence request through SSP sends the message to all leave approvers
configured for a contact, so the message has to be addressed to all approvers,
and not just the first one it finds.

Changed default rule so message is addressed to "Dear Leave Approver",
instead of "Dear, <Manager Name>".