PCHR-2058: Remove code preventing the SSP dashboard to load
The code to rebuild the absence_list view was completely removed, since the view itself will be removed later. It was causing errors becauise it refers to tables created by the hrabsence ext.

One of the job contract view was updated to join with the absence_type table from l&a instead of hrabsence 

Finally, there was a call the HRAbsenceEntitlement, triggered by the My Leave Block that would also make the page to fail to load