PCHR-2058: Avoid API errors when rebuild the absence_activity view
The function to rebuild this view was trying to fetch the custom field
used to store the Absence Comment. This field doesn't exist when the
hrabsence extension is uninstalled, and the CustomGroup.getsingle call
was throwing an error, because getsingle is always expected to find a
single record, but it was finding none.

I also removed the comment field from the view, since it wasn't used by
any of the other views that link to absence_activity