PCHR-2129: Remove Unneeded Commas From Roles in Staff Directory
When a contact had an inactive role and one or more active ones, inactive
roles were being removed from the result set with a hook that altered the
value for the role name, by setting it to null. If the inactive role was the
first role in the result set, this caused the directory to render the list of
roles with a comma inserted before the actual list of roles (eg. ", Role 1,
Role 2, ...").

Fixed by adding filters to the view, to check for roles with start date in the
past or null, and end date in the future or null.