PCHR-2333: Show Document Type on SSP Dashboard
List of documents in SSP dashboard showed an empty type for every document.
This happened because the option values for document types, which in fact are
a special case of activity types, were not being loaded to resolve activity
type id's for their corresponding labels.

Fixed by calling 'getoptions' to Document entity, and merging the result with
original activity types array.