PCHR-2325: Remove custom code from civihr_employee_portal features
On https://github.com/compucorp/civihr-employee-portal/pull/323, a new page was added
to the SSP. To make sure this page would be added to every existing and/or new site,
the civihr_employee_portal feature was updated. However, the update was done manually,
which is now causing two problems:
- The features module now considers this as being overridden
- Some refactoring was added (https://github.com/compucorp/civihr-employee-portal/blob/6979893824369990efdadce82504ac64da8b9b2c/civihr_employee_portal/features/civihr_employee_portal_features/civihr_employee_portal_features.pages_default.inc#L11-L14) with the intent of making the code more readable. This caused a lot more changes to the file than what was really necessary and it's now hard to sync 1.7 with it (due to a big number of merge conflicts)

This PR fixes thoses problems by:
- Adding the new page to the civihr_employee_portal_features.info file, so that the Features module understands it, and take care of this page. This will also help with getting rid of the overridden status.
- Removing all the refactoring from civihr_employee_portal_features.pages_default.inc

Note that this doesn't fix the Overridden status completely. It appears that some other changes had been done manually as well and the feature needs some more updated. It might take some time to fix everything though, so this PR focus only on the fixes required to be able to sync 1.7 with staging.