PCHR-3681: Fix permission for the HR Resources List view
The necessary permission to view this view/page is "edit terms in hr_resource_type"
Where hr_resource_type is a vocabulary created by the civihr_employee_portal
module. Internally, for every vocabulary, Drupal, creates a permission
like 'edit terms in <vocabulary id>'. So, when we export the view, this
ID will be hardcoded in the exported file.

On sites created with the Standard Drupal profile (i.e. all the sites
until now), there's a default "Tags" vocabulary, meaning that when the
"HR Resource Type" is added by the SSP module, the ID of this new
vocabulary will be 2. With the CiviHR profile, the "Tags" vocabulary
isn't created anymore and the the SSP module creates the "HR Resources Type"
vocabulary, it ends up receiving the ID of 1. This commit updates this ID
on the exported view file.