PCHR-3930: Fix upgrader numbers
On https://github.com/compucorp/civihr-employee-portal/pull/532/, an upgrader was added to remove
"Vancacies" from the SSP menu. Originally, this upgrader number was 7036. Before this was merged to
staging, another PR (https://github.com/compucorp/civihr-employee-portal/pull/536) was merged adding
an upgrader with the same number.

This caused a conflict which we tried to fix in https://github.com/compucorp/civihr-employee-portal/pull/532/commits/05cd2c696b25391ef9191e010feded7635d9def6.
When fixing conflicts with upgrader coming from staging, the only correct approach is to change the upgrader in our PR to
a number higher than whatever we have on staging. The reason is that, if we pick a lower number the upgrader will never be executed.
Unfortunately, this is not what happened here. The upgrader was kept with the same name and the number of the one from
staging was increased to 7037.

Finally, https://github.com/compucorp/civihr-employee-portal/pull/532/ was kept open for quite a long time. In the meanwhile,
Another PR (https://github.com/compucorp/civihr-employee-portal/pull/460) was merged to staging adding an upgrader with the number 7037. This time, git didn't
detect the conflict and after https://github.com/compucorp/civihr-employee-portal/pull/532 was merged we ended up with two upgraders
with the same number.

To fix this we:
- Reverted the change in https://github.com/compucorp/civihr-employee-portal/pull/532/commits/05cd2c696b25391ef9191e010feded7635d9def6 and renamed the upgrader that removes the old "Help" menu back to 7036
- Renamed the upgrader added by https://github.com/compucorp/civihr-employee-portal/pull/532 to 7038 (one number higher than the latest we have on staging)