PCHR-4416: Disable progressbars in webforms
The webform module had to be updated to version 7.8 to fix some
compatibility issues with PHP 7.

In this version of the module, progressbars are displayed by default
in single-page forms without a preview page. According to the release
notes (https://www.drupal.org/project/webform/releases/7.x-4.18), we
can run a query to disable this in all the necessary forms. This
was done and then the forms have been exported.

As usual with any webform export, it includes changes that are not
directly related to the updated setting. These changes are related
to new options available in the new version of the module and due to
hardcoded ids in the exported files. The new options can be ignored
and the refresh-node-export-files command, used to refresh the nodes
in existing sites, will take care of fixing the hardcoded ids.