PCHR-4473: Hide the CiviCRM navigation menu on the SSP
With the introduction of the new navigation menu in Civi 5.12.0, the
logic used to display the menu changed a little bit.

The menu is displayed by the CRM.menubar JS component which is only
included on the page according to some certain conditions. One of the
conditions is that the userFrameworkFrontend config is FALSE (i.e the
page being displayed is not a frontend/cms page).

So, to make sure the menu won't be displayed on the SSP, we always
set that config to TRUE when acessing a SSP page.